Moved referral invite into glimmer component
refs https://github.com/TryGhost/Ghost/pull/16187#discussion_r1087258005

- Moved the relevant pieces for the Referral notification into a new `gh-referral-invite` component
