+++ ghost/admin/app/components/gh-nav-menu/footer.hbs
@@ -1,17 +1,12 @@
-<div class="gh-nav-bottom" {{did-insert this.loadCurrentMRR}}>
+<div class="gh-nav-bottom">
     {{#if this.hasThemeErrors}}
     <button type="button" class="gh-footer-toast gh-theme-error-toast" {{on "click" (fn this.openThemeErrors null)}}>
         <span class="gh-footer-toast-title gh-notification-title">Your theme contains errors</span>
         <span class="gh-footer-toast-p">Some functionality on your site may be limited &rarr;</span>
     </button>
-    {{else if this.showReferralInvite}}
-    <a href="https://referrals.ghost.org/?ref=admin" class="gh-referral-toast" target="_blank" rel="noopener noreferrer">
-        <span class="gh-referral-toast-close" {{on "click" (fn this.dismissReferralInvite)}}>&#xd7;</span>
-        <strong>You qualify for our invite-only  referral program.</strong>
-        <p class="gh-footer-toast-p">Refer people to Ghost and earn a <strong>30% share</strong> of the subscription revenue, every single month.</p>
-        <button class="gh-btn gh-btn-black gh-referral-toast-button"><span>Find out more &rarr;</span></button>
-    </a>
     {{/if}}
+
+    <GhReferralInvite @hasThemeErrors={{this.hasThemeErrors}} />
 
     <div class="flex items-center justify-between">
         <div class="pe-all">
